+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.0
+    Added compilation panel position setting
+
+## 0.3.0
+    Added compilation panel
+
 ## 0.2.1
     Fixed highlight bug and warnings bug